Introduce AbstractBackupRepository.
