+++ solr/modules/hdfs/src/java/org/apache/solr/hdfs/backup/repository/HdfsBackupRepository.java
@@ -19,7 +19,6 @@ package org.apache.solr.hdfs.backup.repository;
 
 import java.io.IOException;
 import java.io.OutputStream;
-import java.lang.invoke.MethodHandles;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.Collection;
@@ -37,16 +36,12 @@ import org.apache.solr.common.SolrException;
 import org.apache.solr.common.SolrException.ErrorCode;
 import org.apache.solr.common.util.NamedList;
 import org.apache.solr.core.DirectoryFactory;
-import org.apache.solr.core.backup.repository.BackupRepository;
-import org.apache.solr.core.backup.repository.BackupRepositoryUtil;
+import org.apache.solr.core.backup.repository.AbstractBackupRepository;
 import org.apache.solr.hdfs.HdfsDirectoryFactory;
 import org.apache.solr.hdfs.store.HdfsDirectory;
 import org.apache.solr.hdfs.store.HdfsDirectory.HdfsIndexInput;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-public class HdfsBackupRepository implements BackupRepository {
-  private static final Logger log = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
+public class HdfsBackupRepository extends AbstractBackupRepository {
 
   private static final String HDFS_UMASK_MODE_PARAM = "solr.hdfs.permissions.umask-mode";
   private static final String HDFS_COPY_BUFFER_SIZE_PARAM = "solr.hdfs.buffer.size";
@@ -55,13 +50,11 @@ public class HdfsBackupRepository implements BackupRepository {
   private Configuration hdfsConfig = null;
   private FileSystem fileSystem = null;
   private Path baseHdfsPath = null;
-  private NamedList<?> config = null;
   protected int copyBufferSize = HdfsDirectory.DEFAULT_BUFFER_SIZE;
-  private boolean shouldVerifyChecksum;
 
   @Override
   public void init(NamedList<?> args) {
-    this.config = args;
+    super.init(args);
 
     // Configure the size of the buffer used for copying index files to/from HDFS, if specified.
     if (args.get(HDFS_COPY_BUFFER_SIZE_PARAM) != null) {
@@ -102,13 +95,6 @@ public class HdfsBackupRepository implements BackupRepository {
     } catch (IOException e) {
       throw new SolrException(ErrorCode.SERVER_ERROR, e);
     }
-
-    shouldVerifyChecksum = getBooleanConfig(args, PARAM_VERIFY_CHECKSUM, true);
-  }
-
-  private static boolean getBooleanConfig(NamedList<?> args, String param, boolean defaultValue) {
-    Object value = args.get(param);
-    return value == null ? defaultValue : Boolean.parseBoolean(value.toString());
   }
 
   @Override
@@ -210,17 +196,6 @@ public class HdfsBackupRepository implements BackupRepository {
     try (HdfsDirectory dir =
         new HdfsDirectory(new Path(destDir), NoLockFactory.INSTANCE, hdfsConfig, copyBufferSize)) {
       copyIndexFileFrom(sourceDir, sourceFileName, dir, destFileName);
-    }
-  }
-
-  @Override
-  public void copyIndexFileFrom(
-      Directory sourceDir, String sourceFileName, Directory destDir, String destFileName)
-      throws IOException {
-    if (shouldVerifyChecksum) {
-      BackupRepository.super.copyIndexFileFrom(sourceDir, sourceFileName, destDir, destFileName);
-    } else {
-      BackupRepositoryUtil.copyFileNoChecksum(sourceDir, sourceFileName, destDir, destFileName);
     }
   }
